feat(brs): Add initial (undocumented) version of roca.brs
Our testing API will be shipped with this node module via the `files`
array in package.json, making it accessible to anyone who runs `roca`.

There's still plenty of bugs to shake out and lots of docs to write, but
it at least works for a single level of describe/it!