feat: Add xit and fit with cross-file support
Focusing on a test (via `fit()`) in one file can now exclude non-focused
tests in all other detected files!  Similarly, tests can be skipped
entirely with `xit()`.  This unfortunately requires a pretty big
performance hit: to get a pure-brightscript test runner, we need to call
`Run()` on all test files **twice** - once to discover tests to run (and
whether any tests are focused) and once to actually run the tests after
resolving focus.  That in turn requires those files to be lexed, parsed,
and the resulting ASTs walked two times.  Bummer!

We need to touch all tests twice regardless of what we do so that we can
resolve `fit()` declarations, but there's not much we can do to avoid
re-lexing, re-parsing, etc if we want `roca` tests to also be executable
on-device.  And I _really_ want that.