feat(api): Implement deepEquals test assertion utility (#22)
This includes a 'deepEquals' implementation for making unit test
assertions. Sometimes we want to compare the equality between
associative arrays that have a deep nested structure that a simple
equality check won't do. This implementation performs a recursive check
against data types, compares them, and returns a boolean value if the
two comparative data types are equal or not.