feat: add support for a global setup file (#26)
This adds support for the user to supply a path to a global setup file.
This is helpful for the user because often there are test utilities that
are needed in multiple unit test files. But without a global setup file,
the only way to create those utilities would be to have them in
`source/`, which would mean they get bundled with the actual
application.

This is pretty similar to, for example, Jest's [config
option](https://jestjs.io/docs/en/configuration#globalsetup-string).